+++ pipis/__init__.py
@@ -156,8 +156,22 @@ def update(requirement, name):
 
 @cli.command()
 @click.confirmation_option()
-@click.argument('name', nargs=-1, required=True, type=click.STRING)
-def uninstall(name):
+@click.option('-r', '--requirement')
+@click.argument('name', nargs=-1, type=click.STRING)
+def uninstall(requirement, name):
+    # check presence of args
+    if not (requirement or name):
+        raise click.UsageError('missing arguments/options')
+    # check mutually esclusive args
+    if requirement and name:
+        raise click.UsageError('too much arguments/options')
+    # populate packages list with req file
+    if requirement:
+        try:
+            with open(requirement, 'r') as req:
+                name = map(str.strip, req.readlines())
+        except IOError:
+            raise click.FileError(requirement)
     with click.progressbar(name, label='Removing', item_show_func=_show_package) as packages:
         for package in packages:
             venv_dir = os.path.join(VENV_ROOT_DIR, package)