+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2018-05-18
 ### Added
 - Add `-r/--requirement` for the `remove` command to allow passing `requirements.txt` file.
 - Add `-r/--requirement` for the `update` command to allow passing `requirements.txt` file.
@@ -13,14 +15,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Change output information format fo when package are already/not installed.
 
-## [1.0.1] - 2017-05-17
+## [1.0.1] - 2018-05-17
 ### Fixed
 - Reload `sys.path` and clean it after to avoid issue with "distribution not found".
 
-## [1.0.0] - 2017-05-17
+## [1.0.0] - 2018-05-17
 ### Added
 - Everything.
 
-[Unreleased]: https://gitlab.com/NicolasKAROLAK/pipis/compare/v1.0.1...HEAD
+[Unreleased]: https://gitlab.com/NicolasKAROLAK/pipis/compare/v1.1.0...HEAD
+[1.1.0]: https://gitlab.com/NicolasKAROLAK/pipis/compare/v1.0.1...v1.1.0
 [1.0.1]: https://gitlab.com/NicolasKAROLAK/pipis/compare/v1.0.0...v1.0.1
 [1.0.0]: https://gitlab.com/NicolasKAROLAK/pipis/compare/0c3cc746...v1.0.0